requirements.txt
amqp==2.2.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
cssselect==1.0.3
cssutils==1.0.2
data==0.4
decorator==4.3.0
Django==2.0.4
git+https://github.com/sacovo/django-advanced-filters.git@release/1.0.7
django-braces==1.13.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-easy-select2==1.2.5
django-feedparser==0.2.1
django-grappelli==2.11.1
django-jet==1.0.7
django-markitup==3.0.0
django-multiupload==0.5.2
django-reversion==2.0.13
https://github.com/darklow/django-suit/tarball/v2
django-tinymce==2.7.0
et-xmlfile==1.0.1
feedparser==5.2.1
funcsigs==1.0.2
future==0.16.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.9.2
idna==2.5
ipython-genutils==0.2.0
jdcal==1.3
jsonschema==2.6.0
jupyter-core==4.4.0
keyring==10.4.0
kombu==4.0.2
latex==0.6.4
lxml==3.8.0
Markdown==2.6.7
nbformat==4.4.0
numpy==1.13.1
openpyxl==2.4.1
pandas==0.20.3
plotly==2.5.1
premailer==3.0.1
psycopg2==2.7.1
pycparser==2.17
PySocks==1.6.5
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
requests==2.18.1
SecretStorage==2.3.1
shutilwhich==1.1.0
simplejson==3.6.5
six==1.10.0
tempdir==0.7.1
texttable==0.9.1
traitlets==4.3.2
twilio==6.0.0rc12
urllib3==1.21.1
vine==1.1.3
webencodings==0.5.1
WebOb==1.7.3
django-adminactions
django-colorfield
